Word embeddings batch writer from bytes shared interface
