+++ src/main/scala/com/johnsnowlabs/nlp/training/CoNLL2003NerReader.scala
@@ -6,7 +6,7 @@ import com.johnsnowlabs.ml.crf.{CrfDataset, DatasetMetadata, InstanceLabels, Tex
 import com.johnsnowlabs.nlp.annotators.common.Annotated.PosTaggedSentence
 import com.johnsnowlabs.nlp.annotators.common.{TaggedSentence, TokenPieceEmbeddings, WordpieceEmbeddingsSentence}
 import com.johnsnowlabs.nlp.annotators.ner.crf.{DictionaryFeatures, FeatureGenerator}
-import com.johnsnowlabs.nlp.embeddings.{WordEmbeddingsBinaryIndexer, WordEmbeddingsReader, WordEmbeddingsTextIndexer, WordEmbeddingsReadWriter}
+import com.johnsnowlabs.nlp.embeddings.{WordEmbeddingsBinaryIndexer, WordEmbeddingsReader, WordEmbeddingsTextIndexer, WordEmbeddingsWriter}
 import com.johnsnowlabs.nlp.util.io.{ExternalResource, ReadAs}
 import com.johnsnowlabs.storage.RocksDBConnection
 
@@ -38,9 +38,9 @@ class CoNLL2003NerReader(wordEmbeddingsFile: String,
     if (!new File(fileDb).exists()) {
       embeddingsFormat match {
         case ReadAs.TEXT =>
-          WordEmbeddingsTextIndexer.index(wordEmbeddingsFile, new WordEmbeddingsReadWriter(connection, false, wordEmbeddingsNDims, 5000, 5000))
+          WordEmbeddingsTextIndexer.index(wordEmbeddingsFile, new WordEmbeddingsWriter(connection, false, wordEmbeddingsNDims, 5000, 5000))
         case ReadAs.BINARY =>
-          WordEmbeddingsBinaryIndexer.index(wordEmbeddingsFile, new WordEmbeddingsReadWriter(connection, false, wordEmbeddingsNDims, 5000, 5000))
+          WordEmbeddingsBinaryIndexer.index(wordEmbeddingsFile, new WordEmbeddingsWriter(connection, false, wordEmbeddingsNDims, 5000, 5000))
       }
     }
